Break apart the transaction list styles
