+++ src/modules/FioRequest/components/FioRequestRow.js
@@ -2,17 +2,17 @@
 
 import type { EdgeDenomination } from 'edge-core-js'
 import React, { Component } from 'react'
-import { Image, TouchableHighlight, View } from 'react-native'
+import { Image, StyleSheet, TouchableHighlight, View } from 'react-native'
 import { connect } from 'react-redux'
 
 import fioRequestsIcon from '../../../assets/images/fio/fio_sent_request.png'
 import * as intl from '../../../locales/intl.js'
 import s from '../../../locales/strings'
 import { styles as requestListStyles } from '../../../styles/scenes/FioRequestListStyle'
-import styles from '../../../styles/scenes/TransactionListStyle'
 import THEME from '../../../theme/variables/airbitz'
 import type { State } from '../../../types/reduxTypes'
 import type { FioRequest } from '../../../types/types'
+import { scale } from '../../../util/scaling.js'
 import { getFiatSymbol } from '../../../util/utils'
 import { getDisplayDenomination } from '../../Settings/selectors'
 import T from '../../UI/components/FormattedText/FormattedText.ui.js'
@@ -194,3 +194,97 @@ const mapStateToProps = (state: State, ownProps: OwnProps) => {
 }
 
 export const FioRequestRowConnector = connect(mapStateToProps, {})(FioRequestRow)
+
+const rawStyles = {
+  singleTransaction: {
+    height: scale(80),
+    borderBottomWidth: 1,
+    borderBottomColor: THEME.COLORS.GRAY_3,
+    padding: scale(15),
+    paddingRight: scale(15),
+    paddingLeft: scale(15)
+  },
+  singleTransactionWrap: {
+    backgroundColor: THEME.COLORS.WHITE,
+    flexDirection: 'column',
+    flex: 1
+  },
+  singleDateArea: {
+    backgroundColor: THEME.COLORS.GRAY_4,
+    flex: 3,
+    padding: scale(3),
+    paddingLeft: scale(15),
+    flexDirection: 'row',
+    paddingRight: scale(24)
+  },
+  leftDateArea: {
+    flex: 1
+  },
+  formattedDate: {
+    color: THEME.COLORS.GRAY_2,
+    fontSize: scale(14)
+  },
+
+  transactionInfoWrap: {
+    flex: 1,
+    flexDirection: 'row',
+    height: scale(40)
+  },
+  transactionLeft: {
+    flexDirection: 'row'
+  },
+  transactionLogo: {
+    width: scale(40),
+    height: scale(40),
+    borderRadius: scale(20),
+    marginRight: scale(10)
+  },
+  transactionLeftLogoWrap: {
+    justifyContent: 'center'
+  },
+  transactionPartner: {
+    flex: 1
+  },
+  transactionRight: {
+    flex: 1,
+    width: '100%',
+    height: '100%',
+    flexDirection: 'column',
+    justifyContent: 'center'
+  },
+  transactionTime: {
+    color: THEME.COLORS.SECONDARY
+  },
+  transactionPending: {
+    color: THEME.COLORS.ACCENT_RED
+  },
+  transactionPartialConfirmation: {
+    color: THEME.COLORS.ACCENT_ORANGE
+  },
+  symbol: {
+    fontFamily: THEME.FONTS.SYMBOLS
+  },
+
+  transactionDetailsRow: {
+    flexDirection: 'row',
+    width: '100%'
+  },
+  transactionDetailsRowMargin: {
+    marginBottom: scale(2)
+  },
+  transactionDetailsReceivedTx: {
+    color: THEME.COLORS.TRANSACTION_LIST_RECEIVED_TX
+  },
+  transactionDetailsSentTx: {
+    color: THEME.COLORS.TRANSACTION_LIST_SENT_TX
+  },
+  transactionFiat: {
+    fontSize: 12,
+    color: THEME.COLORS.SECONDARY
+  },
+  transactionPendingTime: {
+    flex: 1,
+    fontSize: 12
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
